# openapi/openapi.bundle.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "校园二手交易平台 API",
    "version": "1.0.0",
    "description": "校园二手交易平台 MVP 的唯一公开 HTTP API 契约。\n\n用户没有固定买家或卖家角色。用户在自己发布的商品中是卖家，\n在自己发起的交易中是买家。推荐算法和邮件通知不在 MVP 范围内。\n\n公开商品列表只展示在售商品，但已认证用户可查看任意状态商品详情。同一买家对\n同一商品只有一个 Trade 购买意向，重复创建请求返回既有意向而不生成第二笔。\n\n跨资源状态变化及并发约束见 `docs/state-machines.md`。\n",
    "contact": {
      "name": "short-term maintainers",
      "url": "https://github.com/KDZZZZZZ/short-term"
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "当前站点 API 根路径"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
    "/auth/register": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "注册学生账号",
        "operationId": "registerUser",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "注册成功并签发访问令牌。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSuccess"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "使用学号和密码登录",
        "operationId": "loginUser",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功并签发访问令牌。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSuccess"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "退出当前登录会话",
        "operationId": "logoutUser",
        "responses": {
          "200": {
            "description": "当前访问令牌已失效或客户端会话已结束。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptySuccess"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/users/me": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "获取当前用户资料",
        "operationId": "getCurrentUser",
        "responses": {
          "200": {
            "description": "当前用户完整资料。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserMeSuccess"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "修改当前用户资料",
        "operationId": "updateCurrentUser",
        "description": "仅修改请求中出现的字段。微信和 QQ 只允许设置或修改为非空值；显式 `null`\n或空字符串返回 `400 / VALIDATION_ERROR`，省略字段表示保持原值。\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "修改后的当前用户资料。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserMeSuccess"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/users/me/password": {
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "修改当前用户密码",
        "operationId": "changeCurrentUserPassword",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "密码修改成功。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptySuccess"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/users/me/products": {
      "get": {
        "tags": [
          "Users",
          "Products"
        ],
        "summary": "获取当前用户发布的商品",
        "operationId": "listCurrentUserProducts",
        "description": "固定按 `created_at DESC, id DESC` 排列。页码分页不提供跨请求快照，并发变更\n可能导致后续页面漂移。每件商品携带 `buyer_review`：该商品收到的买家评价，\n尚无已完成交易的买家发布评价时为 `null`。\n",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProductStatus"
          },
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          }
        ],
        "responses": {
          "200": {
            "description": "当前用户发布的商品分页列表。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyProductPageSuccess"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/users/{userId}/products": {
      "get": {
        "tags": [
          "Users",
          "Products"
        ],
        "summary": "获取其他用户发布的商品",
        "operationId": "listUserProducts",
        "description": "返回该用户在售（`ON_SALE`）和已售出（`SOLD`）的商品，固定按\n`created_at DESC, id DESC` 排列；`RESERVED` 和 `OFF_SHELF` 只有卖家自己\n可以看到。页码分页不提供跨请求快照。用户不存在返回\n`404 / RESOURCE_NOT_FOUND`。已售出商品的买家评价通过商品详情展示。\n",
        "parameters": [
          {
            "$ref": "#/components/parameters/UserId"
          },
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          }
        ],
        "responses": {
          "200": {
            "description": "该用户公开可见的商品分页列表。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductPageSuccess"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/products": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "浏览和搜索在售商品",
        "operationId": "listProducts",
        "description": "仅返回 `ON_SALE` 商品，固定按 `created_at DESC, id DESC` 排列。页码分页\n不提供跨请求快照，并发发布可能导致后续页面漂移。\n",
        "parameters": [
          {
            "$ref": "#/components/parameters/Keyword"
          },
          {
            "$ref": "#/components/parameters/Category"
          },
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          }
        ],
        "responses": {
          "200": {
            "description": "在售商品分页列表。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductPageSuccess"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "tags": [
          "Products"
        ],
        "summary": "发布商品",
        "operationId": "createProduct",
        "description": "当前用户微信或 QQ 至少填写一项，最多同时上传三张图片。图片按 multipart 中的\n出现顺序获得连续 `sort_order`，第一张图片成为封面。整个请求体不超过\n16 MiB，单张图片不超过 5 MiB。\n",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreateRequest"
              },
              "encoding": {
                "images": {
                  "style": "form",
                  "explode": true,
                  "contentType": "image/jpeg, image/png, image/webp"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "商品已发布，初始状态为 `ON_SALE`。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductSuccess"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/products/{productId}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ProductId"
        }
      ],
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "获取商品详情和卖家联系方式",
        "operationId": "getProduct",
        "description": "对所有已认证用户返回任意状态商品的详情；`ON_SALE`、`OFF_SHELF`、\n`RESERVED` 和 `SOLD` 的可见性相同。`404` 只表示商品不存在。响应返回\n卖家昵称、微信和 QQ，但不返回卖家学号。`buyer_review` 在商品已完成\n交易且买家已发布评价时为该评价（此时商品为 `SOLD`），否则为 `null`。\n",
        "responses": {
          "200": {
            "description": "商品详情。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductSuccess"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "patch": {
        "tags": [
          "Products"
        ],
        "summary": "修改自己发布的商品",
        "operationId": "updateProduct",
        "description": "仅商品卖家可操作，且不能借此接口修改商品状态。只有 `ON_SALE` 或\n`OFF_SHELF` 且不存在 `PENDING` 购买意向的商品可以修改；存在 `PENDING`\n意向时返回 `409 / TRADE_STATE_CONFLICT`，`RESERVED` 或 `SOLD` 返回\n`409 / PRODUCT_STATE_CONFLICT`。修改与创建意向必须共同锁定 Product，保证\n新意向的价格快照和商品内容来自同一已提交版本。\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "修改后的商品详情。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductSuccess"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/products/{productId}/images": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ProductId"
        }
      ],
      "post": {
        "tags": [
          "Products"
        ],
        "summary": "为自己的商品补充图片",
        "operationId": "addProductImages",
        "description": "只有商品卖家可以操作，且商品必须为 `ON_SALE` 或 `OFF_SHELF`；\n商品还不得存在 `PENDING` 购买意向。存在 `PENDING` 意向时返回\n`409 / TRADE_STATE_CONFLICT`，`RESERVED` 或 `SOLD` 返回\n`409 / PRODUCT_STATE_CONFLICT`。新增后商品图片总数不能超过三张；新图片按\n上传顺序追加到现有图片之后并取得连续 `sort_order`，仅当此前没有图片时第一张\n新图片成为封面。超过三张时返回 `400 / IMAGE_LIMIT_EXCEEDED`。整个请求体不超过\n16 MiB，单张图片不超过 5 MiB。\n",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/ImageUploadRequest"
              },
              "encoding": {
                "images": {
                  "style": "form",
                  "explode": true,
                  "contentType": "image/jpeg, image/png, image/webp"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "商品当前全部图片。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductImageListSuccess"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/products/{productId}/images/{imageId}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ProductId"
        },
        {
          "$ref": "#/components/parameters/ImageId"
        }
      ],
      "delete": {
        "tags": [
          "Products"
        ],
        "summary": "删除自己商品的一张图片",
        "operationId": "deleteProductImage",
        "description": "只有商品卖家可以操作，且商品必须为 `ON_SALE` 或 `OFF_SHELF`；\n商品还不得存在 `PENDING` 购买意向。存在 `PENDING` 意向时返回\n`409 / TRADE_STATE_CONFLICT`，`RESERVED` 或 `SOLD` 返回\n`409 / PRODUCT_STATE_CONFLICT`。删除后其余图片保持相对顺序并重新编号为连续的\n`1..N`；若删除封面，新的第一张图片成为封面，没有剩余图片时 `cover_url` 为\n`null`。\n",
        "responses": {
          "200": {
            "description": "图片已删除。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptySuccess"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/products/{productId}/off-shelf": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ProductId"
        }
      ],
      "post": {
        "tags": [
          "Products"
        ],
        "summary": "下架自己的商品",
        "operationId": "offShelfProduct",
        "description": "仅商品卖家可执行 `ON_SALE -> OFF_SHELF`，且商品不得存在任何 `PENDING`\n购买意向。存在 `PENDING` 意向时必须先逐笔拒绝，由该意向与下架操作共同锁定\nProduct 后返回 `409 / TRADE_STATE_CONFLICT`，不得出现已下架商品仍有\n`PENDING` 意向的提交结果。\n",
        "responses": {
          "200": {
            "description": "下架后的商品详情。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductSuccess"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/products/{productId}/relist": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ProductId"
        }
      ],
      "post": {
        "tags": [
          "Products"
        ],
        "summary": "重新上架自己的商品",
        "operationId": "relistProduct",
        "description": "仅允许 `OFF_SHELF -> ON_SALE`。",
        "responses": {
          "200": {
            "description": "重新上架后的商品详情。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductSuccess"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/products/{productId}/comments": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ProductId"
        }
      ],
      "get": {
        "tags": [
          "Comments"
        ],
        "summary": "查看商品的用户评论",
        "operationId": "listProductComments",
        "description": "返回该商品收到的全部用户评论，固定按 `created_at DESC, id DESC` 排列。\n评论与商品状态无关，`OFF_SHELF`、`RESERVED` 和 `SOLD` 商品的评论同样可见；\n`404` 只表示商品不存在。页码分页不提供跨请求快照，并发评论可能导致\n后续页面漂移。\n",
        "parameters": [
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          }
        ],
        "responses": {
          "200": {
            "description": "商品用户评论分页列表。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentPageSuccess"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "tags": [
          "Comments"
        ],
        "summary": "发布用户评论",
        "operationId": "createProductComment",
        "description": "任何已认证用户都可以对任何现存商品发布评论：不要求购买过该商品，不限\n商品状态，用户也可以评论自己发布的商品。每个用户对同一商品可以发布\n多条评论。评论发布后不可修改、不可删除。评论发布不改变商品或交易状态。\n商品不存在返回 `404 / RESOURCE_NOT_FOUND`。\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "评论已发布。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentSuccess"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/favorites": {
      "get": {
        "tags": [
          "Favorites"
        ],
        "summary": "获取当前用户的收藏列表",
        "operationId": "listFavorites",
        "description": "已下架、已预留或已售商品仍可返回，并携带当前商品状态。固定按\n`favorited_at DESC, product.id DESC` 排列；页码分页不提供跨请求快照。\n",
        "parameters": [
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          }
        ],
        "responses": {
          "200": {
            "description": "收藏商品分页列表。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FavoritePageSuccess"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/favorites/{productId}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ProductId"
        }
      ],
      "put": {
        "tags": [
          "Favorites"
        ],
        "summary": "收藏商品",
        "operationId": "addFavorite",
        "description": "接口幂等；任意状态的现存商品均可收藏，包括 `OFF_SHELF`、`RESERVED` 和\n`SOLD`。用户不能收藏自己发布的商品；该情况固定返回\n`409 / SELF_ACTION_NOT_ALLOWED`。`400` 仅用于请求格式或参数值无效。\n",
        "responses": {
          "200": {
            "description": "商品处于已收藏状态。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptySuccess"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "tags": [
          "Favorites"
        ],
        "summary": "取消收藏商品",
        "operationId": "removeFavorite",
        "description": "接口幂等，收藏记录不存在时也返回成功。",
        "responses": {
          "200": {
            "description": "商品处于未收藏状态。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptySuccess"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/products/{productId}/conversations": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ProductId"
        }
      ],
      "post": {
        "tags": [
          "Conversations"
        ],
        "summary": "创建或获取商品会话",
        "operationId": "getOrCreateProductConversation",
        "description": "当前用户作为买家，商品发布者作为卖家。相同商品、买家和卖家的会话只存在一个；\n重复调用返回原会话。用户不能针对自己的商品创建会话，该情况固定返回\n`409 / SELF_ACTION_NOT_ALLOWED`；`400` 仅用于请求格式或参数值无效。\n",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "responses": {
          "200": {
            "description": "新建或已经存在的会话。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationSuccess"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "登录或注册接口签发的访问令牌。"
      }
    },
    "schemas": {
...
      "RegisterRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "student_no",
          "password"
        ],
        "properties": {
          "student_no": {
            "$ref": "#/components/schemas/StudentNumber"
          },
          "password": {
            "$ref": "#/components/schemas/PasswordInput"
          },
          "nickname": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Nickname"
              }
            ],
            "description": "省略时服务端使用“校园用户”；默认值不得从学号派生。",
            "default": "校园用户"
          },
          "wechat": {
            "$ref": "#/components/schemas/Wechat"
          },
          "qq": {
            "$ref": "#/components/schemas/QQ"
          }
        }
      },
...
      "AuthSuccess": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SuccessBase"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/AuthData"
              }
            }
          }
        ]
      },
...
      "LoginRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "student_no",
          "password"
        ],
        "properties": {
          "student_no": {
            "$ref": "#/components/schemas/StudentNumber"
          },
          "password": {
            "$ref": "#/components/schemas/PasswordInput"
          }
        }
      },
...
    },
...
  }
}